Added 2D histogram of Jet Charge vs W p_T, added Condor scripts for running on cluster.
2D histogram looks at correlation between high p_T jets and jet
charge, should be linearly correlated because of the kappa weighting
(higher the pt of the jet, bettter determined the charge is)

Condor scripts will steer the high statistics jobs on atlxxx machines.
They're here for easy scripting and pushing and pulling to remote
site.  I may branch them into their own github repository eventually,
but for now they'll come as part of the project.