Switched from 15->30 min pt in all generator configs, added min pt and max eta cuts to Sherpa, added truth level plots to analysis code.
Extract the truth level quark from the event and plot relevant variables [3/3]
directions to see which ones Jet Charge is sensitive to vs not.
The truth level quark for a jet is defined as the highest p_T quark
within a cone of the Jet.
 - [X] \Delta R between quark and jet (0 \leq \Delta R \leq 0.6)
 - [X] Split out plots based on PdgID {quark-,gluon-}{jet-pt,jet-charge}
 - [X] Make PdgID plot (set gluon to 0 to avoid awkward plot range)

Tweaked Pythia8 generator files to allow switching between PDF sets.